msm: kgsl: Write TSB registers after turning on power
Due to changes in gpu power on sequence in recent gpus,
these register writes are misplaced in the sense that
the gpu power and clocks is not really on at the time
of these register writes. Hence, these writes are dropped.
So perform these register writes when we are sure that
GPU power and clocks are up.

Change-Id: Ie1e4fd1a08e283ebc0f4110f6d4fe20f61b6ef3c
Signed-off-by: Harshdeep Dhatt <hdhatt@codeaurora.org>
Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>
Signed-off-by: celtare21 <celtare21@gmail.com>